+++ pkg/models/shared/wiseconfig.go
@@ -2,13 +2,31 @@
 
 package shared
 
-type WiseConfig struct {
-	APIKey string `json:"apiKey"`
+type DummyPayConfig struct {
+	Directory string `json:"directory"`
+	// The frequency at which the connector will create new payment objects in the directory
+	FileGenerationPeriod *string `json:"fileGenerationPeriod,omitempty"`
+	// The frequency at which the connector will try to fetch new payment objects from the directory
+	FilePollingPeriod *string `json:"filePollingPeriod,omitempty"`
 }
 
-func (o *WiseConfig) GetAPIKey() string {
+func (o *DummyPayConfig) GetDirectory() string {
 	if o == nil {
 		return ""
 	}
-	return o.APIKey
+	return o.Directory
+}
+
+func (o *DummyPayConfig) GetFileGenerationPeriod() *string {
+	if o == nil {
+		return nil
+	}
+	return o.FileGenerationPeriod
+}
+
+func (o *DummyPayConfig) GetFilePollingPeriod() *string {
+	if o == nil {
+		return nil
+	}
+	return o.FilePollingPeriod
 }